ec2 logs performance monitoring/requirements.txt
kafka-python
pandas
python-dotenv
pyodbc